+++ src/movie_list.h
@@ -11,9 +11,9 @@ MovieList *movie_list_create(void);
 void movie_list_destroy(MovieList *list);
 
 /* Operations */
-size_t movie_list_size(const MovieList *list);
-const Movie *movie_list_get(const MovieList *list, size_t index); //read-only
-Movie *movie_list_get_mut(MovieList *list, size_t index); //mutable
+int movie_list_size(const MovieList *list);
+const Movie *movie_list_get(const MovieList *list, int index); //read-only
+Movie *movie_list_get_mut(MovieList *list, int index); //mutable
 void movie_list_add(MovieList *list, Movie *movie);
 void movie_list_print_table(const MovieList *list);
